# Conflicts:
#	script/uePlug/Doodle/Source/Doodle/Private/Doodle/AiArrayGeneration.cpp
#	script/uePlug/Doodle/Source/DoodleEditor/Private/Doodle/CreateCharacter/CharacterEditorViewport.cpp
#	script/uePlug/Doodle/Source/DoodleEditor/Private/Doodle/CreateCharacter/CoreData/DoodleCreateCharacterConfig.cpp
#	script/uePlug/Doodle/Source/DoodleEditor/Private/Doodle/CreateCharacter/CoreData/DoodleCreateCharacterConfig.h
#	script/uePlug/Doodle/Source/DoodleEditor/Private/Doodle/CreateCharacter/CreateCharacterCurveEditor.cpp
#	script/uePlug/Doodle/Source/DoodleEditor/Private/Doodle/CreateCharacter/CreateCharacterSliderController.cpp
#	script/uePlug/Doodle/Source/DoodleEditor/Private/Doodle/CreateCharacter/CreateCharacterTree.cpp
#	script/uePlug/Doodle/Source/DoodleEditor/Private/DoodleEditor.cpp
#	script/uePlug/Doodle/Source/doodleUI/Private/doodleUI.cpp

+++ src/doodle_lib/exe_warp/maya_exe.cpp
@@ -4,6 +4,12 @@
 
 #include "maya_exe.h"
 
+#include "doodle_core/configure/config.h"
+#include "doodle_core/core/doodle_lib.h"
+#include "doodle_core/core/file_sys.h"
+#include "doodle_core/core/global_function.h"
+#include "doodle_core/logger/logger.h"
+#include <doodle_core/configure/config.h>
 #include <doodle_core/core/core_set.h>
 #include <doodle_core/core/program_info.h>
 #include <doodle_core/lib_warp/boost_fmt_error.h>
@@ -15,19 +21,40 @@
 
 #include <doodle_lib/core/filesystem_extend.h>
 
+#include "boost/signals2/connection.hpp"
 #include <boost/asio.hpp>
+#include <boost/asio/bind_cancellation_slot.hpp>
 #include <boost/asio/high_resolution_timer.hpp>
 
+#include "exe_warp/maya_exe.h"
+#include <filesystem>
+#include <fmt/core.h>
 #include <stack>
+#include <string>
+#include <winnt.h>
+#include <winreg/WinReg.hpp>
 // #include <type_traits>
 
+#include <boost/asio/readable_pipe.hpp>
+#include <boost/core/ignore_unused.hpp>
+#include <boost/locale/encoding_utf.hpp>
 #include <boost/process.hpp>
+#include <boost/process/environment.hpp>
 #include <boost/process/extend.hpp>
+// #include <boost/process/v2/environment.hpp>
+// #include <boost/process/v2/execute.hpp>
+// #include <boost/process/v2/process.hpp>
+// #include <boost/process/v2/stdio.hpp>
 #ifdef _WIN32
 #include <boost/process/windows.hpp>
 #elif defined __linux__
 #include <boost/process/posix.hpp>
 #endif
+// #define BOOST_PROCESS_V2_SEPARATE_COMPILATION
+
+// #include <boost/process/v2.hpp>
+// #include <boost/process/v2/src.hpp>
+
 namespace doodle {
 namespace {
 // 致命错误。尝试在 C:/Users/ADMINI~1/AppData/Local/Temp/Administrator.20210906.2300.ma 中保存
@@ -56,7 +83,10 @@ class run_maya : public std::enable_shared_from_this<run_maya> {
   entt::handle mag_attr{};
   FSys::path file_path_attr{};
 
-  std::string run_script_attr{};
+  nlohmann::json run_script_attr{};
+  std::string run_script_attr_key{};
+  FSys::path program_path{};
+  FSys::path maya_program_path{};
 
   boost::process::async_pipe out_attr{g_io_context()};
   boost::process::async_pipe err_attr{g_io_context()};
@@ -68,24 +98,26 @@ class run_maya : public std::enable_shared_from_this<run_maya> {
   std::shared_ptr<std::function<void(boost::system::error_code)>> call_attr{};
 
   boost::asio::high_resolution_timer timer_attr{g_io_context()};
+  boost::signals2::scoped_connection cancel_attr{};
 
   run_maya()          = default;
   virtual ~run_maya() = default;
 
   void run() {
-    auto l_maya     = core_set::get_set().maya_path() / "mayapy.exe";
-    auto l_tmp_file = FSys::write_tmp_file("maya", run_script_attr, ".py");
-    auto &&l_msg    = mag_attr.get<process_message>();
+    if (program_path.empty()) throw doodle_error{"没有找到maya路径 (例如 C:/Program Files/Autodesk/Maya2019/bin})"};
+
+    auto l_path  = FSys::write_tmp_file("maya", run_script_attr.dump(), ".json");
+
+    auto &&l_msg = mag_attr.get<process_message>();
     l_msg.set_state(l_msg.run);
-    l_msg.set_name(file_path_attr.filename().generic_string());
-    l_msg.message(fmt::format("开始写入配置文件 {} \n", l_tmp_file), l_msg.warning);
-    l_msg.aborted_function = [this]() {
+    l_msg.message(fmt::format("开始写入配置文件 {} \n", l_path), l_msg.warning);
+    cancel_attr = l_msg.aborted_sig.connect([this]() {
       auto &&l_msg = mag_attr.get<process_message>();
       l_msg.set_state(l_msg.fail);
       l_msg.message("进程被主动结束\n");
       child_attr.terminate();
       cancel();
-    };
+    });
 
     timer_attr.expires_from_now(chrono::seconds{core_set::get_set().timeout});
     timer_attr.async_wait([this](boost::system::error_code in_code) {
@@ -99,10 +131,32 @@ class run_maya : public std::enable_shared_from_this<run_maya> {
       }
     });
 
+    boost::process::environment l_eve = boost::this_process::environment();
+    l_eve["MAYA_LOCATION"]            = maya_program_path.generic_string();
+    l_eve["Path"] += (maya_program_path / "bin").generic_string();
+    l_eve["Path"] += program_path.parent_path().generic_string();
+    l_eve["Path"] += core_set::get_set().program_location().generic_string();
+
+    // boost::process::v2::process_environment l_env{
+    //     std::unordered_map<boost::process::v2::environment::key, boost::process::v2::environment::value>{
+    //         {"MAYA_LOCATION"s, program_path.parent_path().parent_path().generic_string()},
+    //         {"PATH"s, program_path.parent_path().generic_string()}}};
+
+    // boost::asio::readable_pipe l_out_pipe{g_io_context()};
+    // boost::asio::readable_pipe l_err_pipe{g_io_context()};
+    // boost::process::v2::process_stdio l_io{nullptr, l_out_pipe, l_err_pipe};
+    // boost::process::v2::process l_process{
+    //     g_io_context(), program_path, {fmt::format("--{}={}", run_script_attr_key, l_path)}, l_env, l_io};
+    // boost::asio::cancellation_signal sig{};
+    // boost::process::v2::async_execute(
+    //     boost::process::v2::process{
+    //         g_io_context(), program_path, {fmt::format("--{}={}", run_script_attr_key, l_path)}, l_env, l_io},
+    //     boost::asio::bind_cancellation_slot(sig, [](boost::system::error_code ec, int exit_code) {})
+    // );
     child_attr = boost::process::child{
         g_io_context(),
-        boost::process::exe  = l_maya,
-        boost::process::args = l_tmp_file.generic_wstring(),
+        boost::process::exe  = program_path.generic_string(),
+        boost::process::args = fmt::format("--{}={}", run_script_attr_key, l_path),
         boost::process::std_out > out_attr,
         boost::process::std_err > err_attr,
         boost::process::on_exit =
@@ -113,7 +167,8 @@ class run_maya : public std::enable_shared_from_this<run_maya> {
               l_msg.message(fmt::format("退出代码 {}", in_exit));
               (*call_attr)(in_error_code);
             },
-        boost::process::windows::hide
+        boost::process::windows::hide,
+        l_eve
 
     };
 
@@ -138,7 +193,7 @@ class run_maya : public std::enable_shared_from_this<run_maya> {
             read_out();
           } else {
             out_attr.close();
-            DOODLE_LOG_ERROR(in_code);
+            DOODLE_LOG_ERROR(in_code.what());
           }
         }
     );
@@ -175,7 +230,7 @@ class run_maya : public std::enable_shared_from_this<run_maya> {
             read_err();
           } else {
             err_attr.close();
-            DOODLE_LOG_ERROR(in_code);
+            DOODLE_LOG_ERROR(in_code.what());
           }
         }
     );
@@ -187,103 +242,6 @@ class run_maya : public std::enable_shared_from_this<run_maya> {
   }
 };
 
-std::string qcloth_arg::to_str() const {
-  nlohmann::json l_json{};
-  l_json = *this;
-  return fmt::format(
-      R"(# -*- coding: utf-8 -*-
-import maya_fun_tool
-k_f =  maya_fun_tool.open_file()
-k_f.config_ = """{}"""
-k_f()
-)",
-      l_json.dump()
-  );
-}
-
-std::string export_fbx_arg::to_str() const {
-  nlohmann::json l_json{};
-  l_json = *this;
-  return fmt::format(
-      R"(# -*- coding: utf-8 -*-
-import maya_fun_tool
-k_f =  maya_fun_tool.open_file()
-k_f.config_ = """{}"""
-k_f()
-)",
-      l_json.dump()
-  );
-}
-
-std::string replace_file_arg::to_str() const {
-  auto l_list_entt = save_handle | ranges::views::transform([](const entt::handle &in) -> std::string {
-                       nlohmann::json l_j{};
-                       entt_tool::save_comm<assets_file, redirection_path_info>(in, l_j);
-                       return fmt::format(R"(cmds.doodle_add_entt(josndata="""{}"""))", l_j.dump());
-                     }) |
-                     ranges::to_vector;
-
-  return fmt::format(
-      R"(# -*- coding: utf-8 -*-
-from maya import cmds
-import pymel.core
-import maya.standalone
-maya.standalone.initialize()
-
-cmds.file(force=True, new=True)
-doodle_plug = "doodle_maya_" + str(cmds.about(api=True))[0:4]
-cmds.loadPlugin(doodle_plug)
-
-l_file_path = "{0}"
-project_path = "{1}"
-work_path = "{2}"
-cmds.workspace(work_path, openWorkspace=1)
-
-cmds.file(l_file_path, open=True)
-cmds.doodle_load_project(project=project_path)
-
-{3}
-
-cmds.doodle_create_ref_file(allload=True)
-cmds.doodle_replace_rig_file()
-cmds.doodle_comm_file_save()
-quit()
-)",
-      file_path, project_, find_maya_work(file_path).generic_string(), fmt::join(l_list_entt, "\n")
-  );
-}
-
-std::string clear_file_arg::to_str() const {
-  auto l_save_file_path =
-      file_path.parent_path() / "fbx" / file_path.filename().replace_extension(save_file_extension_attr);
-  return fmt::format(
-      R"(# -*- coding: utf-8 -*-
-from maya import cmds
-import pymel.core
-import maya.standalone
-maya.standalone.initialize()
-
-cmds.file(force=True, new=True)
-
-l_file_path = "{0}"
-save_file_path = "{1}"
-project_path = "{2}"
-work_path = "{3}"
-doodle_plug = "doodle_maya_" + str(cmds.about(api=True))[0:4]
-cmds.loadPlugin(doodle_plug)
-cmds.workspace(work_path, openWorkspace=1)
-
-cmds.doodle_load_project(project=project_path)
-cmds.file(l_file_path, open=True)
-
-cmds.doodle_clear_scene(err_4=True)
-cmds.doodle_comm_file_save(filepath=save_file_path)
-quit()
-)",
-      file_path, l_save_file_path, project_, find_maya_work(file_path).generic_string()
-  );
-}
-
 }  // namespace maya_exe_ns
 
 class maya_exe::impl {
@@ -292,19 +250,49 @@ class maya_exe::impl {
   std::vector<std::shared_ptr<maya_exe_ns::run_maya>> run_attr{};
 
   std::atomic_char16_t run_size_attr{};
+  FSys::path run_path{};
 };
 
 maya_exe::maya_exe() : p_i(std::make_unique<impl>()) {}
 
-void maya_exe::add_maya_fun_tool() {
-  auto l_file_py = cmrc::DoodleLibResource::get_filesystem().open("resource/maya_fun_tool.py");
-  FSys::write_tmp_file(
-      "maya", std::string{l_file_py.begin(), boost::numeric_cast<std::size_t>(l_file_py.size())}, ".py",
-      "maya_fun_tool", std::ios::out | std::ios::binary
-  );
+FSys::path maya_exe::find_maya_path() const {
+  boost::ignore_unused(this);
+  auto l_key_str = fmt::format(LR"(SOFTWARE\Autodesk\Maya\{}\Setup\InstallPath)", core_set::get_set().maya_version);
+  DOODLE_LOG_INFO("开始寻找在注册表中寻找maya");
+  winreg::RegKey l_key{};
+  l_key.Open(HKEY_LOCAL_MACHINE, l_key_str, KEY_QUERY_VALUE | KEY_WOW64_64KEY);
+  auto l_maya_path = l_key.GetStringValue(LR"(MAYA_INSTALL_LOCATION)");
+  return l_maya_path;
 }
+
+void maya_exe::install_maya_exe() {
+  try {
+    auto l_maya_path   = find_maya_path();
+
+    auto l_target_path = FSys::get_cache_path() / fmt::format("maya_{}", core_set::get_set().maya_version) /
+                         version::build_info::get().version_str;
+    const auto l_run_name = fmt::format("doodle_maya_exe_{}.exe", core_set::get_set().maya_version);
+    p_i->run_path         = l_target_path / l_run_name;
+
+    if (!FSys::exists(l_target_path)) FSys::create_directories(l_target_path);
+
+    if (!FSys::exists(l_target_path / "ShadeFragment")) {
+      FSys::copy(l_maya_path / "bin" / "ShadeFragment", l_target_path / "ShadeFragment", FSys::copy_options::recursive);
+    }
+    if (!FSys::exists(l_target_path / "ScriptFragment")) {
+      FSys::copy(
+          l_maya_path / "bin" / "ScriptFragment", l_target_path / "ScriptFragment", FSys::copy_options::recursive
+      );
+    }
+    if (!FSys::exists(p_i->run_path)) {
+      FSys::copy(core_set::get_set().program_location() / l_run_name, l_target_path / l_run_name);
+    }
+  } catch (const winreg::RegException &in_err) {
+    DOODLE_LOG_WARN("在注册表中寻找maya失败,错误信息: {}", boost::diagnostic_information(in_err));
+  }
+}
+
 void maya_exe::notify_run() {
-  add_maya_fun_tool();
   if (!doodle_lib::Get().ctx().get<program_info>().stop_attr())
     while (p_i->run_size_attr < core_set::get_set().p_max_thread && !p_i->run_process_arg_attr.empty()) {
       auto l_run = p_i->run_process_arg_attr.top();
@@ -324,17 +312,19 @@ void maya_exe::notify_run() {
   }
 }
 void maya_exe::queue_up(
-    const entt::handle &in_msg, const std::string &in_string, const std::shared_ptr<call_fun_type> &in_call_fun,
-    const FSys::path &in_run_path
+    const entt::handle &in_msg, const std::string_view &in_key, const nlohmann::json &in_string,
+    const std::shared_ptr<call_fun_type> &in_call_fun, const FSys::path &in_run_path
 ) {
-  DOODLE_CHICK(
-      core_set::get_set().has_maya(), doodle_error{"没有找到maya路径 (例如 C:/Program Files/Autodesk/Maya2019/bin})"}
-  );
-  auto l_run             = p_i->run_process_arg_attr.emplace(std::make_shared<maya_exe_ns::run_maya>());
-  l_run->mag_attr        = in_msg;
-  l_run->run_script_attr = in_string;
-  l_run->file_path_attr  = in_run_path;
-  auto &&l_msg           = in_msg.get<process_message>();
+  install_maya_exe();
+
+  auto l_run                 = p_i->run_process_arg_attr.emplace(std::make_shared<maya_exe_ns::run_maya>());
+  l_run->mag_attr            = in_msg;
+  l_run->run_script_attr_key = in_key;
+  l_run->run_script_attr     = in_string;
+  l_run->file_path_attr      = in_run_path;
+  l_run->program_path        = p_i->run_path;
+  l_run->maya_program_path   = find_maya_path();
+  auto &&l_msg               = in_msg.get<process_message>();
   l_msg.set_name(in_run_path.filename().generic_string());
   l_run->call_attr =
       std::make_shared<call_fun_type>([in_call_fun, this, in_msg](const boost::system::error_code &in_code) {
